+++ lib/gamefic/standard.rb
@@ -19,6 +19,7 @@ module Gamefic
     require 'gamefic/standard/openable'
     require 'gamefic/standard/lockable'
     require 'gamefic/standard/direction'
+    require 'gamefic/standard/implicit_taking'
     require 'gamefic/standard/entities'
     require 'gamefic/standard/actions'
     require 'gamefic/standard/introduction'