+++ lib/gamefic/standard/actions/give.rb
@@ -16,9 +16,10 @@ module Gamefic
         end
 
         respond :give, available(Character), available do |actor, _character, gift|
-          actor.execute :take, gift if gift.parent != actor
+          actor.proceed if actor.have_or_take(gift)
+          # actor.execute :take, gift if gift.parent != actor
 
-          actor.proceed if gift.parent == actor
+          # actor.proceed if gift.parent == actor
         end
 
         respond :give, Thing do |actor, thing|