+++ lib/gamefic/standard/actions/lock.rb
@@ -32,8 +32,7 @@ module Gamefic
         end
 
         respond :lock, available(Lockable, proc(&:has_lock_key?)), available do |actor, thing, key|
-          actor.execute :take, key if key.parent != actor
-          actor.proceed if key.parent == actor
+          actor.proceed if actor.have_or_take(key)
         end
 
         interpret "lock :container with :key", "lock :container :key"